requirements.txt
qwen-asr
torchaudio
fastapi
uvicorn
python-multipart